requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
d6tcollect==1.0.6
d6tflow==0.2.6
decorator==4.4.2
docutils==0.19
flatbuffers==23.3.3
frozenlist==1.3.3
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
h5py==3.8.0
httplib2==0.22.0
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.2.0
jax==0.4.8
keras==2.12.0
libclang==16.0.0
lockfile==0.12.2
luigi==3.2.1
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.0.4
moviepy==1.0.3
multidict==6.0.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.4
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
proglog==0.1.10
protobuf==4.22.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyobjc==9.0.1
pyobjc-core==9.0.1
pyobjc-framework-Accounts==9.0.1
pyobjc-framework-AddressBook==9.0.1
pyobjc-framework-AdSupport==9.0.1
pyobjc-framework-AppleScriptKit==9.0.1
pyobjc-framework-AppleScriptObjC==9.0.1
pyobjc-framework-ApplicationServices==9.0.1
pyobjc-framework-AudioVideoBridging==9.0.1
pyobjc-framework-AuthenticationServices==9.0.1
pyobjc-framework-AutomaticAssessmentConfiguration==9.0.1
pyobjc-framework-Automator==9.0.1
pyobjc-framework-AVFoundation==9.0.1
pyobjc-framework-AVKit==9.0.1
pyobjc-framework-BusinessChat==9.0.1
pyobjc-framework-CalendarStore==9.0.1
pyobjc-framework-CFNetwork==9.0.1
pyobjc-framework-CloudKit==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-Collaboration==9.0.1
pyobjc-framework-ColorSync==9.0.1
pyobjc-framework-Contacts==9.0.1
pyobjc-framework-ContactsUI==9.0.1
pyobjc-framework-CoreAudio==9.0.1
pyobjc-framework-CoreAudioKit==9.0.1
pyobjc-framework-CoreBluetooth==9.0.1
pyobjc-framework-CoreData==9.0.1
pyobjc-framework-CoreHaptics==9.0.1
pyobjc-framework-CoreLocation==9.0.1
pyobjc-framework-CoreMedia==9.0.1
pyobjc-framework-CoreMediaIO==9.0.1
pyobjc-framework-CoreMIDI==9.0.1
pyobjc-framework-CoreML==9.0.1
pyobjc-framework-CoreMotion==9.0.1
pyobjc-framework-CoreServices==9.0.1
pyobjc-framework-CoreSpotlight==9.0.1
pyobjc-framework-CoreText==9.0.1
pyobjc-framework-CoreWLAN==9.0.1
pyobjc-framework-CryptoTokenKit==9.0.1
pyobjc-framework-DeviceCheck==9.0.1
pyobjc-framework-DictionaryServices==9.0.1
pyobjc-framework-DiscRecording==9.0.1
pyobjc-framework-DiscRecordingUI==9.0.1
pyobjc-framework-DiskArbitration==9.0.1
pyobjc-framework-DVDPlayback==9.0.1
pyobjc-framework-EventKit==9.0.1
pyobjc-framework-ExceptionHandling==9.0.1
pyobjc-framework-ExecutionPolicy==9.0.1
pyobjc-framework-ExternalAccessory==9.0.1
pyobjc-framework-FileProvider==9.0.1
pyobjc-framework-FileProviderUI==9.0.1
pyobjc-framework-FinderSync==9.0.1
pyobjc-framework-FSEvents==9.0.1
pyobjc-framework-GameCenter==9.0.1
pyobjc-framework-GameController==9.0.1
pyobjc-framework-GameKit==9.0.1
pyobjc-framework-GameplayKit==9.0.1
pyobjc-framework-ImageCaptureCore==9.0.1
pyobjc-framework-IMServicePlugIn==9.0.1
pyobjc-framework-InputMethodKit==9.0.1
pyobjc-framework-InstallerPlugins==9.0.1
pyobjc-framework-InstantMessage==9.0.1
pyobjc-framework-Intents==9.0.1
pyobjc-framework-IOSurface==9.0.1
pyobjc-framework-iTunesLibrary==9.0.1
pyobjc-framework-LatentSemanticMapping==9.0.1
pyobjc-framework-LaunchServices==9.0.1
pyobjc-framework-libdispatch==9.0.1
pyobjc-framework-LinkPresentation==9.0.1
pyobjc-framework-LocalAuthentication==9.0.1
pyobjc-framework-MapKit==9.0.1
pyobjc-framework-MediaAccessibility==9.0.1
pyobjc-framework-MediaLibrary==9.0.1
pyobjc-framework-MediaPlayer==9.0.1
pyobjc-framework-MediaToolbox==9.0.1
pyobjc-framework-Metal==9.0.1
pyobjc-framework-MetalKit==9.0.1
pyobjc-framework-MetalPerformanceShaders==9.0.1
pyobjc-framework-ModelIO==9.0.1
pyobjc-framework-MultipeerConnectivity==9.0.1
pyobjc-framework-NaturalLanguage==9.0.1
pyobjc-framework-NetFS==9.0.1
pyobjc-framework-Network==9.0.1
pyobjc-framework-NetworkExtension==9.0.1
pyobjc-framework-NotificationCenter==9.0.1
pyobjc-framework-OpenDirectory==9.0.1
pyobjc-framework-OSAKit==9.0.1
pyobjc-framework-OSLog==9.0.1
pyobjc-framework-PencilKit==9.0.1
pyobjc-framework-Photos==9.0.1
pyobjc-framework-PhotosUI==9.0.1
pyobjc-framework-PreferencePanes==9.0.1
pyobjc-framework-PushKit==9.0.1
pyobjc-framework-Quartz==9.0.1
pyobjc-framework-QuickLookThumbnailing==9.0.1
pyobjc-framework-SafariServices==9.0.1
pyobjc-framework-SceneKit==9.0.1
pyobjc-framework-ScreenSaver==9.0.1
pyobjc-framework-ScriptingBridge==9.0.1
pyobjc-framework-SearchKit==9.0.1
pyobjc-framework-Security==9.0.1
pyobjc-framework-SecurityFoundation==9.0.1
pyobjc-framework-SecurityInterface==9.0.1
pyobjc-framework-ServiceManagement==9.0.1
pyobjc-framework-Social==9.0.1
pyobjc-framework-SoundAnalysis==9.0.1
pyobjc-framework-Speech==9.0.1
pyobjc-framework-SpriteKit==9.0.1
pyobjc-framework-StoreKit==9.0.1
pyobjc-framework-SyncServices==9.0.1
pyobjc-framework-SystemConfiguration==9.0.1
pyobjc-framework-SystemExtensions==9.0.1
pyobjc-framework-UserNotifications==9.0.1
pyobjc-framework-VideoSubscriberAccount==9.0.1
pyobjc-framework-VideoToolbox==9.0.1
pyobjc-framework-Vision==9.0.1
pyobjc-framework-WebKit==9.0.1
pyparsing==3.0.9
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
zipp==3.15.0
